Add export of NUMBER_SEQUENCE_KEYS to constants file
