+++ src/database/utilities/constants.js
@@ -5,3 +5,12 @@
  */
 
 export const MILLISECONDS_PER_DAY = 24 * 60 * 60 * 1000;
+
+export const NUMBER_SEQUENCE_KEYS = {
+  CUSTOMER_INVOICE_NUMBER: 'customer_invoice_serial_number',
+  INVENTORY_ADJUSTMENT_SERIAL_NUMBER: 'inventory_adjustment_serial_number',
+  REQUISITION_SERIAL_NUMBER: 'requisition_serial_number',
+  REQUISITION_REQUESTER_REFERENCE: 'requisition_requester_reference',
+  STOCKTAKE_SERIAL_NUMBER: 'stocktake_serial_number',
+  SUPPLIER_INVOICE_NUMBER: 'supplier_invoice_serial_number',
+};
