+++ src/database/utilities/index.js
@@ -3,13 +3,7 @@
  * Sustainable Solutions (NZ) Ltd. 2019
  */
 
-export {
-  createRecord,
-  getNextNumber,
-  getNumberSequence,
-  reuseNumber,
-  NUMBER_SEQUENCE_KEYS,
-} from './createRecord';
+export { createRecord, getNextNumber, getNumberSequence, reuseNumber } from './createRecord';
 export { deleteRecord } from './deleteRecord';
 export { getTotal, addBatchToParent, millisecondsToDays } from './utilities';
-export { MILLISECONDS_PER_DAY } from './constants';
+export { MILLISECONDS_PER_DAY, NUMBER_SEQUENCE_KEYS } from './constants';
